Introduce PreviewMode concept
This CL builds upon change I1a73a1.

This refactoring introduces a concept of PreviewMode alongside a
StateMachine. This will make handling transitions from one state to
another easier. It also has the benefit of simplifying some checks.

Before we would rely on setting variables and cleaning up when exiting a
specific mode. With the preview modes,we can rely on the modes storing
the necessary information. We can also centralise the setting up and
cleaning up when transitioning from one mode to another in the preview
representation.

One follow-up purpose of this refactoring will be to extract some of the
common logic so that it can be used by other previews.

Bug: 228294269
Test: bazel test //tools/adt/idea/compose-designer/...
Change-Id: I5646228e084193f0e114f0dccdca0df7308c5879