Fix Gradle Failures reporting
Before this change failure details were collected by IssueCheckers
that recognize the failure. But this collected value was never
sent to the usage log, unless it stayed in memory and next sync
accidentally triggers the codepath to send this value. This makes
currently gathered data on sync failure details pretty useless.

Refactor code to make sure that after each sync failure the collected
value is sent. Also clean any possible stale value on sync start.
SyncFailureUsageReporter is an application service and projects are
distinguished by the path. This is because IssueCheckers that detect
failure do not have Project object but only project path. Using paths
means that we don't need an expensive operation of finding the
right project at the moment of error processing (see code removal in
BuildIssueUtil.kt).

Regarding tests, this feature should be tested as part of integration
tests for error checkers. Currently, Jdk8RequiredErrorTest.kt was
adjusted to check that the event is logged after this error is triggered
in sync. This covers just enough to see that this part of infrastructure
works. Other checkers tests seem to not check this which will be fixed
in the further work.

Bug: 289903923
Test: affected infrastructure covered in Jdk8RequiredErrorTest.kt
Change-Id: I7f2784479b020d5a9e125dea66599ed79d04a80f