IDEA-323476 Update default value for BUILD_OUTPUT_DOWNLOADS_INFORMATION
Modified the accessor for the `BUILD_OUTPUT_DOWNLOADS_INFORMATION` StudioFlag from a simple property get to a method call `isBuildOutputShowsDownloadInfo()`. In this case, it checks if the Flag is overridden, if not it defaults to `IdeInfo.getInstance().isAndroidStudio()`. This ensures that the feature is not enabled by default in IDEA while maintaining a possibility to be changed by a user.
 Move flag check to an outer level, to prevent unnecessary code execution when the flag is disabled.

Merge-request: IJ-MR-109445
Merged-by: Dmitry Yudin <Dmitry.Yudin@jetbrains.com>

GitOrigin-RevId: aff1eb9e77c9bedce7fde14c70980db3b13072c4
(cherry picked from commit 29237bd01c4019657b1abd504252644db820b7ab)
Change-Id: I2c5609a5bef0b072b9c22843045d9dc0e80e8ae0