Implement highlighting validation for KTS files on Android projects
The new tests represent an addition to the platform located
under GradleBuildFileHighlightingTest, and validates android
specific projects that are not covered by the existing ones.

This is important, since in Android projects the Android SDK is used
to define the module SDK that may result in unresolved symbols since
this jdk.table entity doesn't contain any link to the JDK classpath root.

Bug: 289494258
Bug: 287967703
Test: GradleKtsResolveSymbolsTest
Change-Id: I93472e012e64cd70dedd4cc33635df82ee1e418e