Display UI Check mode errors in their own tab
When in UI Check mode, the Visual lint issues are displayed in a new tab
in the Problems panel. This tab is closable. Otherwise the tab persists
even after UI Check mode is stopped.
If starting several UI Checks in several Compose files, there is a tab
per file, which displays the errors coming from that file.
When selecting an issue, it will be highlighted as before, but switching
to a different tab will make the highlighting disappear.

Bug: 219402820
Test: tests added
Change-Id: I8a60efb1ce152759a3589eca8882f03de6eb0ef6