Fixes bug when reselecting current device in Preview Picker
When a @Preview function had the device pararameter set using the "id"
property of the device spec, e.g. "id:pixel", setting the corresponding
device using the preview picker (e.g. "Pixel") would cause the device
argument to be rewritten as a string of the FQN of the device constant,
e.g. "androidx.compose.ui.tooling.preview.Devices.PIXEL".

The issue was happening because PsiDropDownModel sets the selected value
using the EnumSupport object when the picker is still loading the other
values of the model. This makes the selectedValue point to Devices.PIXEL
instead of "id:pixel". In setSelectedItem, however, we were not updating
the selectedValue if the **resolved** value of the newly selected value
is equal to the current one. Since "id:pixel" and Devices.PIXEL resolve
to the same thing, selectedValue remained as Devices.PIXEL. This causes
a problem later because the ComboboxModel thinks these values are not
the same and try to update "id:pixel" to "Devices.PIXEL".

This issue is easily fixed by updating selectedValue according to the
user selection, so the ComboboxModel check will be happy.

Fixes: 290582306
Test: PsiPropertyDropDownTest
Change-Id: I4a44392c178017d719ec4977940c928e131f86c3