Wait for refresh when starting Interactive preview
It was possible to finish the mode transition before the
invokeOnCompletion was called. This would mean that it would appear like
the transition to the interactive preview mode was finished when there
were still some setup to perform.

By actively waiting for the refresh, we can ensure the remaining setup
is performed before the mode switching is over.

Bug: N/A
Test: bazel test //tools/adt/idea/compose-designer/...
Change-Id: Icaa609c1cb219a274f0b727806cc9ae3d148be33