[Dsl] Strengthen project dependency tests
Check for the final form of the build file Dsl when modifying the name
of the dependency.  Whether the fact that the `path` argument is moved
to the end is truly "expected" is arguable, but it is at least
syntactically correct, and generic to the MapMethodTransform that
implements the Dsl operations.

Bug: none filed
Test: strengthened existing tests
Change-Id: I8d4ccc4e1c3e22ead9439193a5ccd1804d4b5073