Migrate to Gradle identity path for build path
Use absolute Gradle path to refer to builds instead
of build names (that are not unique if there are
multiple nested builds with the same name).

This change modifies BuildMapModelBuilder to use
identityPath.path instead of identityPath.name,
and this is used to map the location of Gradle build
to its path.

The IDE import code has been modified to handle this,
and ProjectInfo.buildId is patched to always contain
the absolute path, allowing us to match project dependencies
with IDE modules.

Bug: 282119683
Test: existing sync tests
Change-Id: I032e19a32cf2a40307eaa5ba454c2eae4df0222e