Add Dagger test for target element navigation
The existing tests validate that the correct gutter icons show up in the
correct places, and do indeed click on them; but none validate that the
cursor moves where it's supposed to. The single test case added is
sufficient to cover the various Dagger component types, since all use
the same underlying platform code to navigate to the corresponding
PsiElement and other tests validate that the correct PsiElements are
being displayed.

Bug: n/a
Test: test code
Change-Id: I86c1587208aef30419207cc05b5aef5f16be5dcb