Add a DataContext key for PreviewModeManager
... to not depend on ComposePreviewManager directly when setting a mode
in the preview actions.

This change is a first step, as there are still references to the actual
preview instance, which is a compose preview representation.
Furthermore, we rely on compose specific text/information to update the
action presentation at the moment. Those links will be broken in
following changes.

Bug: 228294269
Test: existing
Change-Id: I17cd1ee71e17cab092b2296a7a97a76861f18688