Remove RunConfigurationChecker
This was added in 2018 to help mitigate an issue from Android Studio 3.0
to 3.1. The issue was that Gradle aware make task was missing from run
configuration loaded from 3.0.

All the required run configuration should be created with this task added.

This only impacts users opening any project previously opened in
3.0 and earlier, and if they do recreating the run configurations or
manually adding the task is a low friction workaround. I think it makes
sense to remove.

Bug: 76403139
Test: N/A
Change-Id: Ia92e9fdf7c23a8012d712e932166649f4165ef6c