Expose HeapSnapshotStatistics outside of its package
Before this change, it was possible to easily perform memory usage
performance tests using the e2e integration test framework, but not
in other tests.

This change adds a way to collect memory usage information using the
HeapSnapshotTraverseService in a more opened way, that could be use,
for example, in any perfgate test of any tool.

Bug: 286520840
Bug: 279604596
Test: N/A
Change-Id: Ibeb34adc86138d65e1bbde1510bc2d43fe6408aa