LE: Add a way to read legacy modes
We removed enum entries from LiveEditTriggerMode when we
were adding the ON_HOTKEY trigger. That's probably not a
good idea given we rely on intellij's state persisting APIs.

When IntelliJ attempts to restore a enum value that no longer
exists, it restores it as a null value to a non-nullable Kotlin
function which causes the NPE (and changing the default mode).

The best way to remedy that is to restore the old enum types
and migrate to the newer version as needed. The drawback would
be LiveEditTriggerMode having two extra modes that are unused
which is the price we pay for a better user experience.

Fixes: 282990657
Fixes: 280516816

Test: LiveEditApplicationConfigurationTest
Change-Id: I99e02f23395649369ceaa21d3da977b9363e470c