Don't filter out deprecated devices on @Preview validation
We were filtering out deprecated devices on the validation of the device
parameter of @Preview annotations. As a result, setting a valid value,
e.g. "id:Nexus 5" would trigger a syntax error and the preview picker
would not show up.

This issue can be fixed by not filtering out deprecated devices on the
validation.

Fixes: 291553716
Test: PreviewAnnotationCheckTest#deprecatedDevicesShouldBeValid
Change-Id: Ic010c41243d8f6681618346fce541d1b11bd0dc6