Object depth calculation for extended reports implemented
(part 1)
Now on the second traverse besides attributing objects to components we
will also calculate the length of the minimal path from the GC root to
each object. This information will be used in the heap traverse
collecting shortest GC root paths for instances of nominated classes
(and retained disposed objects).

Optimal distance from the GC root is stored in ExtendedHeapTraverseNode
in a separate map, because it will only be calculated for extended
reports.

Bug: 280408311
Test: None
Change-Id: I6366e03ebbd9b8d8b854fe7911c177a294f405a6