Fix deadlock in PreviewRefreshManager
The deadlock was happening between the UI thread and the requestsLock
in the PreviewRefreshManager.

The two conflicting paths generating the deadlock were:

1. The PreviewRefreshManager collects a request from its requestsFlow
in the background thread. When processing the request, it acquires
the requestsLock and starts a new refresh job without releasing the
lock. When setting up the refresh job, it is usually necessary to
use the UI thread.

2. A client of the PreviewRefreshManager (e.g. Compose Preview) needs
to request a new refresh due to some change. This new refresh request
is sent created using PreviewRefreshManager.requestRefresh, which needs
to acquire the requestsLock for adding the new request to the "queue".
Then, when this request is created from the UI thread, the deadlock may
occur.

This change solves the problem by releasing the requestsLock before
creating a new refresh job in the PreviewRefreshManager. This is
done using CoroutineStart.LAZY, which makes a job not to start until
indicated to do so by calling job.start() or job.join().

Fixes: 291792172
Test: PreviewRefreshManagerTest
Change-Id: I609a1c941e96f14164d753cc36b85726351d1d9c