Revert "Set StringResourceTableModel model off EDT"
This reverts commit f6c318eb009aef744c01a2c8e4ecc87bef3b079b.

The change caused Swing methods to be invoked off the EDT, leading to
deadlocks seen in b/291567167 and b/291600637. The revert moves loading
of the model back to the EDT, while retaining a small UI change to
improve the loading screen.

Additionally, hiding the UI while it reloads led to initial focus on the
editor being incorrect. Reverting this brings back some slightly broken
UI after the user clicks the refresh button, but that's a less unsightly
bug than having the focus be incorrect on every load.

The original change here was an optimization that wasn't necessary to
fix the freeze in b/286457185, so I'm comfortable just moving this back.
There's a little more happening here on the EDT than I'd like, but it's
not really possible to address without rewriting pieces of the
ResourceManager. So I'm inclined to leave this as it is now unless we
start to see actual freezes coming out of it.

Fixes: 291567167
Fixes: 291600637
Fixes: 290097248
Bug: 286457185
Test: n/a
Change-Id: Idf7bb9df9d38a6ec23a346713292beded2d13f4e