Enable/Disable Tool Windows Dynamically
This cl restores the idea from ag/21404138 but using the correct
test for `is Android Project`.

The Tool Windows in this cl used `ToolWindowFactory.isApplicable()` to
enable/disable themselves. This does not work well in IntelliJ because
`isApplicable()` is only called once when the project is loaded. But in
IntelliJ, a non Android project can become an Android project by adding
an Android module. With the current code, this project will not have
any Android Tool Windows until it is reloaded.

The ``<library.toolWindow>` extension point solves this by reevaluating
the applicability of a Tool Window whenever the project structure
changes.

Bug: 277937502
Test: Added
Change-Id: I9cc83e913035d865493d2fee93ac010a47e6eb8f