Root paths for nominated classes added to extended reports
(part 2)
Now extended memory usage reports will contain detailed information
about instances of the nominated classes owned by exceeded components.

Nominated classes are:
1) Top 3 classes with the highest number of instances;
2) Top 3 classes with the highest total size of instances;
3) Top 3 component root classes with the highest number of instances.

For each of the nominated classes we collect the trees of shorted paths
to roots retaining instances of the class. This information is intended
to help in the detection of the leakage paths and potential
memory optimizations.

Bug: 280408311
Test: HeapAnalyzerTest#testExtendedReportNominatedClassesTree
Change-Id: I4b7170b33a0ffbdede669bf4c4ccd05e35f05b71