[uitest-framework] Remove BlockReloading
The call to blockReloadingProjectOnExternalChanges() was originally
added in Change If84060bbf in 2015 to work around a race condition
in IntelliJ. The original change did not link to a bug or YouTrack
issue, so it unclear whether the race condition has been fixed
upstream. However, 8 years is long enough to asssume bit rot.

Also, this code is no longer compatible with IntelliJ 2023.2,
because StoreReloadManager is now a project-level service.

Change-Id: I74a19b6dcf2e1a13009e5e2073967c4132c1cc46