Refactor material icon support
To accomodate for supporting Material Symbols. Material Symbols are a
new set of icons many of which duplicate Material Icons. Because of
that, we can not rely on the icon name as a unique key anymore.
In this change we move to using the name to use the full metadata class.

In a future CL we will move the URL to point the Material Symbols one.

Bug: 293255004
Test: //tools/adt/idea/android-material
Change-Id: Ia01be7c8f583c01ee3520e59c41bba6f74538c37