Always consider the new project version available
To allow testing of the upgrade assistant when using injected
repositories with a staged google maven snapshot containing
the newer AGP.

This partially reverts the behaviour change in
https://cs.android.com/android-studio/platform/tools/base/+/a21e842810659affab05a151f1d03fad478a5483
but only for the staged build case.

For nightly builds and for local development run from the IDE
where AGP has not been built, the latest known AGP version
will not be included as an option to update to.

Fixes: 292203730
Test: Manually verified, added additional unit test.
Change-Id: I5bd95b56b954a6e2072817cc965383ff6875b7f4