Fix flaky CommonPreviewModeManagerTest
The test was flaky due to the initialisation of lastMode happening in
the coroutine scope. In the flaky cases, it would not execute before
the test started calling setMode, resulting in lastMode not being the
expected value of PreviewMode.Default, but PreviewMode.Switching.

There was a call to yield() in the test to allow time for the coroutine
to execute, but it seems like that was not foolproof. This change moves
the initialisation of lastMode outside the coroutine and into the init
block to ensure it gets initialised with PreviewMode.Default.

Fixes: 293501939
Test: bazel test //tools/adt/idea/preview-designer:\
  intellij.android.preview-designer.tests_tests\
  --test_filter=CommonPreviewModeManagerTest\
  --config=remote --runs_per_test=500
Change-Id: Ia1b139c5fe16b7655ce23964b4d0ffb972e05730