Avoid deadlock in ProguardR8UseScopeEnlarger
This enlarger can be called on the EDT when in response to a direct user
action, such as Highlight Usages (Ctrl+Shift+F7). If indexes aren't yet
initialized, this leads to a deadlock where the logic on the EDT is
waiting for initialization to finish.

I unfortunately was not able to reproduce the scenario in a test
environment, since the platform does not seem to have any granular
controls to get the indexes into this state.

Fixes: 292099949
Test: Manually verified
Change-Id: I1448901effe46e9d40f7241d2a955fbe94b447e2