Make the fetch of memory recording options lazy
To make the recording options reusable for task based ux, they have
been lazily made and retrieved.

This is done as in order to utilize a recording option, it must be
already populated in the recording options model. Thus, we must
maintain the instance of the recording option. This must be done in
task based ux as we need to fetch the recording option to start or
stop the recording.

Bug: 277798936
Test: added
Change-Id: I9dbe7472c584f9af21d6bbf0ddc04776d60a30f8