Fix model dumper for KotlinGradleModel.compilerArgumentsBySourceSet
Platform 232 removed cachedCompilerArgumentsBySourceSet field from
the KotlinGradleModel that represented the compiled arguments using
K2JVMCompilerArguments, end up handled by CommonCompilerArguments
custom dumper that displayed the arguments as properties but also
filtered some of them that weren't considered relevant.

Since we started using the compilerArgumentsBySourceSet the arguments
are represented via List<String>, ignoring the custom dumper. For
this reason a SpecializedDumper was created for this field to be able
to keep same structure we had before on the snapshots.

Bug: 293473692
Test: Existing snapshots
Change-Id: Ieba08a0b1e415654510902cd268246867a054509