Enable desugaring when dexing during evaluate expression.
When the class files produced by the Kotlin compiler contains
instructions that are not supported by the eval4j interpreter,
the code is compiled to dex and loaded in a class loader on
device.

When using the Kotlin JVM_IR backend, the Kotlin compiler
now generates lambdas using invoke-dynamic instructions and
the lambda metafactory. That requires us to perform desugaring
during dexing. Unfortunately, we have desugaring turned off
for our evaluate expression dexing. Most likely in an attempt
to make this dexing as fast as possible?

This chagne enables desugaring and also sets the min api level
higher. We only use dexing if we are on Android O+. Setting the
min api level to 26 (Android O) will limit the amount of work
desugaring has to do.

Bug: 294485700
Test: Manual

Change-Id: Ibf76bd8792e222f5e523a662bbdd4d0123a4c793