Re-enable Java tests for TypeDefCompletionContributorTest
This change makes a first stab at fixing
TypeDefCompletionContributorTest. Specifically, it simplifies how we
find and add completion results. Instead of just decorating existing
results, we add our own and then filter out existing results. This
creates a more consisten experience instead of relying on the
underlying system to suggest results (which it sometimes fails to do).

The next step will be to figure out why resolving the TypeDefs in Kotlin doesn't always work in tests, and then re-enabling the test for Kotlin as well.

Bug: 294309761
Test: TypeDefCompletionContributorTest.kt
Change-Id: Iea2fca3daa10661a18c53787f1e99a3e4daabe66