Fix concurrency issue in ModelBuilderService parameters
Before this change, there is only a single instance of
ModelBuilderContext per service instance and that is used to contain
the model builder parameters. But multiple `buildAll` actions can  be
run concurrently resulting in model builders getting the paramaters for
other model builders. This change moves the parameter out of the
ModelBuilderContext and to the method parameters to avoid this issue.

closes https://github.com/JetBrains/intellij-community/pull/2535

(cherry pick of commit 6d665022a93b9546b0eefc2d1a3ec4414b9d8591)

Change-Id: I6d665022a93b9546b0eefc2d1a3ec4414b9d8591