Open settings panel when opening Essentials Mode settings
When clicking on the bottom-right Essentials Mode icon, and then on
"Settings Page", we were showing a popup with only the Essentials Mode
settings. Instead, we should properly open the settings panel and
navigate to the essentials portion of it.

Fixes: 293831366
Test: N/A. Manually verified.
Change-Id: Ie51d87a9abab3583cc0f8c77e64f75621aa246d0
(cherry picked from commit ea3d453f640a739629fc0ceabaaca477fc58f19e)