Trigger a highlight pass when changing essentials mode state
When enabling/disablng essentials mode via the File -> Essentials Mode
menu action, we were not triggering a highlight on the selected files.
As a result, some highlighting and gutter icons (e.g. deploy preview)
that shouldn't be present in essentials mode were showing until the
user either modified the file or changed tabs back and forth.

This change fixes this issue by triggering a single highlight pass when
enabling/disabling essentials mode, so the editor state is correctly
immediately after the action is triggered.

Fixes: 294086553
Test: N/A. Manually verified.
Change-Id: Iccbfef95abec02a23ee1448643958eeac244ee68