Update device details on status changes
Currently, available storage and power is scanned once when creating the
DeviceInfoPanel. Normally, this is fine, as these properties do not
change very quickly in most cases, and the DeviceInfoPanel is recreated
every time the user selects a different device in the Device Manager.

However, if the user happens to create the DeviceInfoPanel while the
device is still booting, it will not get a value for these properties
(until the panel is recreated, perhaps by momentarily selecting a
different device).

Similarly, the ABI property of FTL devices is not present in the
template; after reservation, but before connection, we observe the
DirectAccessDeviceHandle in a disconnected state, before having read the
ABI from the system properties.

With this change, we update power and storage when the device's
connection status changes, and we update device properties whenever they
change.

Bug: 294346976
Test: Updated
Change-Id: I000c8fa0ce916f05dd4a31a4dd4d9742611d15b3