Execute Everything in EDT
We filter the tool windows so we only process ones that are not loaded.
But we do this filtering in the background and then switch to the EDT,
and by then, they could be loaded already.

So, we move the filtering into EDT as well.
Since registering is only done in EDT, this removes the race condition.

Fixes: 295201351
Test: NA
Change-Id: I48527c09ddf48ee7249097eb895c254ebdc8f9e2