Fix stat for DeployerException with non-none ResolutionAction
If we have a resolution action for deployer exception we show a custom
error bubble. To prevent platform for showing error twice we used
CantRunException.CustomProcessedCantRunException(). As a side affect of
using CantRunException.CustomProcessedCantRunException() we didn't
propagate error.id to statistics, for that we need to use
AndroidExecutionException.

This cl fixes it.

Stop adding duplication of error message to custom bubbleError, only
leave information there about resolution action.

Fixes: 295328564
Test: DeployAndHandleErrorKtTest
Change-Id: I55217c0390828371b49cc8f50afc014cefe2a060