[SDK Index] Fix text on PSD for two or more issues
... in the same library.

When two or more issues are present for the same libary version, in lint
we show a message that says that "one or more issues" are present and in
the PSD a similar message should appear but instead a single type
appears.

The reason is that in PSD to know if two or more issue types are
present, a boolean for each type is used and then those booleans are put
in a collection and at the end we simply count how many of those
boolenas are true. Unfortunately this only works if the collection
allows repeated values and the current implementation uses a set.

This CL changes setOf(<list of booleans>) with listOf(<list of
booleans>) and adds tests to confirm the expected message is created in
the reported issue.

Bug: 284994378
Test: Added PsAnalyzerDaemonKtTest
Change-Id: I0228d0b4e88fb1bcf20e342f44d68eff204cd83c