Fix background of TwoLineLabel in Paired Devices
TwoLineLabel implements TableComponent so that it can make its second
line a slightly lighter color. The default behavior is to call
applyColors on all components recursively. However, the overridden
implementation here fails to set the colors for the panel itself; it
applies colors only to its children.

This normally doesn't matter, because the panel is not opaque, so its
colors do nothing. However, the table cell component currently under the
mouse is made opaque and painted over other components to deal with
overflow.

This doesn't happen on the devices table because TwoLineLabel is wrapped
by DeviceNamePanel, which has the default behavior and is the component
made opaque when hovered.

Bug: 294349380
Test: updated
Change-Id: Ibbdbc1e644907d3ce25e5040f16a668c07d14363