Enable session creation when setting same process
In the Task-Based UX, when the user launches a task and thus
begins a new session, it is done via setting the process.
This can be done with the same process consecutively. However,
in the current code, beginning a session inside the setProcess
method will only occur if the process sent into the method is
different than the current process.

This change disables this check for the same process if the
Task-Based UX is enabled, allowing for beginning a new session
even if you call setProcess with the same process consecutively.

Bug: 277797528
Test: added
Change-Id: Iaf1cdb8e65d6348edaf0d28580504ad031655d97