Listen to Essentials Mode state in Design Tools settings
The Design Tools settings panel contains a few elements whose enable and
visible states should match Android Studio Essentials Mode state. This
was working fine as long as the user didn't change the Essentials Mode
state using the settings panel, pressed "Apply" and then navigated to
the Design Tools settings panel. In this case, we'd still show the panel
considering the Essentials Mode state at the moment the user opened the
settings panel.

This behavior is fixed in this change by creating an observable property
that listens to change on the Essentials Mode status. The UI elements of
the design tools settings panel rely on the observable to update their
state accordingly.

Fixes: 295577721
Test: N/A. Manually verified.
Change-Id: Ia5bf7d1141efb4e8ceca0831bf870e46622ec8a6