Disable MEMORY_CAPTURE session change listener
...so that the MainMemoryProfilerStage is not automatically
entered when a memory capture is imported to the Profiler.
This disabling is done only when the Task-Based UX flag is
enabled.

This is done because in the Task-Based UX, importing will not
trigger any stage change, but rather just adds the import to
a list of past recordings. Then a user can choose to take action
on the imported capture/past recording via a task.

Note: this Change is a companion change to Change
Iceb41ae5fa6e09f53b34cd3a255ae4879cacec04.

Bug: 277798530
Test: added
Change-Id: If581dd1284487c76cf844d3e15de982c23d4efdf