Link via signatures in live edit.
This ensures that the backend will only see kotlin basic types
from the stdlib that is used for the current compilation.

The default for JVM Kotlin compilation is to link via descriptors.
That works because for any ordinary compilation it is guaranteed
that there will only be one descriptor for a primitive type.

In the IDE, the setup is more complicated and in particular
for libraries, we see multiple descriptors for the same primitive
type from different versions of the Kotlin standard library.
Therefore, linking via descriptors ends up with multiple versions
of the same types which breaks compilation.

Linking via signatures avoids this problem. Any descriptor that
corresponds to a kotlin primitive type will be mapped via
the symbol table to the representation of that type used by
the current compilation.

Bug: 256957527
Test: Manual
Change-Id: I65ae50787ce75f2236f728b9e1cd7d426e111a76