Prevent recompute of all nodes + rectanges on selection
Previously, every time the user selected a new node in
the HTreeChart, it would trigger a recompute of all nodes
and their respective rectanges. This is unnecessary on
only a selection change, as it only requires a re-render.

Thus, now a selection change only triggers a nulling out
of the canvas, which will trigger only a re-render.

Bug: 281850040
Test: manual
Change-Id: Ib86372fa57831c39fff8775a77505f4a1639f6e3