Precompute name to nodes mapping
The two heaviest computations done post-selection of
a capture node are the top k occurrences of the node
in the thread and the analysis stats for the node.

The bulk of each computation is actually filtering for
the nodes with a matching name from all descendants of
the root node. Because both computations are done for
the root node for the thread, we now precompute a
mapping of the names to a list of nodes with the
respective name for each thread track model. Precomputing
these values saves a massive amount of time now when
computing and displaying selected capture node details.

Bug: 281850040
Test: added + updated
Change-Id: I420c23f9c418e50ba44ef517752c01f570d191b9