Make project should assemble all modules
There is no easy way for users to trigger
assembling of all modules in the IDE. This change
modifies behavior of the "Make project" action and
it triggers assemble for all modules in the IDE.

Previously, we had complex logic to build only
application modules and avoid building libraries
that are in the application's dependency graph.
However, Gradle already makes sure no task is
executed twice, and for such a heavy action
this optimization seems neglible.

Bug: 69829055
Test: MakeGradleProjectActionTest
Change-Id: I2b36b8b05f80e10869b036f7a94e4f30f45a2e28