[Sync] Notification for programmatic catalog definition
At (post-)sync time, check whether there are programmatic catalog
entries defined in Gradle settings files, and if so provide a
notification that Android Studio does not support this mode, and that
the user should consider moving the catalog entries to a Toml file.

This notification obeys the modern NotificationGroup standards, and
consequently the user has direct control over whether and how to see
that notification (in the Appearance & Behavior part of the Settings
dialog, and directly from the context menu of the notification): they
can reschedule the notification or disable it project- or
application-wide.

Fixes: 295069426
Test: added
Change-Id: Ie3dbaae42391a6be21928e593f95ad0b4d89c8b0