[Dagger] Add a central DaggerClasses object.
This object will contain DaggerClass objects, each of which contains
fully-qualified class IDs for all Dagger-related typenames. Class IDs,
unlike their string counterparts, are explicitly divided into separate
package and relative-class components.

This is in preparation for porting Dagger to K2, which depends on
ClassIds in its own lookups. Since ClassIds can be converted to strings
easily, but not vice-versa, ClassIds are the best canonical
representation of a known class's name.

Bug: 278637936,279139016
Test: Treehugger

Change-Id: Idf91d9f4c5e27b4885ef4c315f76b2592ad992eb