Remove Layout Inspector UI on tab close
When a RD tab with LI is closed, we should clean up LI by removing the
injected UI and all the listeners that were registered.
This prevents leaking memory when a tab is closed.

While removing the injected UI is not critical, as the tab's UI will be
destroyed anyway, removing the listeners on Layout Inspector is.
Not removing them will prevent the SelectedTabState instance from
being collected.

Bug: 298200058
Test: LayoutInspectorManagerTest
Change-Id: I1429618c65f110143e1d5b8f30ff112efb044e2b