Add ProfilerTaskHandler class
The ProfilerTaskHandler serves as the base class for
all task handlers. It enforces implementation of what
to do on task start, stop, and how to load a task.
Moreover, it also enforces implementation for a method
to create the respective task arguments, when to know if
the task is defined as terminated, what to return for
the task name, and whether the task supports a specific
artifact.

Bug: 277797528
Test: added
Change-Id: I5fe58e17461cdf0ab0cedb47bcd885fce3acf37e