Only generate test R classes for test modules
ProjectLightResourceClassService generates light R classes for modules
by creating a bunch of variants and then paring them down later. It
currently generates both "main" and "test" variants for any given
module, which is a relic of the system prior to using linked modules.
Now that there are separate modules for the main and test source sets,
we should no longer be returning the test R class.

In practical terms, the impact of this change is limited and affects
only a few apis:

`getLightRClassesDefinedByModule`: This is only called by
NamespacesRefactoringsUtil. The caller is updated in this change to
explicitly request R classes for the test module as well as the main
module.

`getAllLightRClasses`: This method already loops through all facets,
meaning it would have been generated test module R classes twice (once
for the main module, and once for the test module). This reduces the
duplication, but will still return all the necessary modules.

`getLightRClasses`: This method would get the extra test R classes when
calling `classes.all` for the main module, but then would filter them
out because they didn't have the required qualified name. By no longer
returning the test R class, we've just removed a class from
consideration that wouldn't have been returned anyway.

Bug: 292491619
Test: TestRClassesTest
Change-Id: Ic1ce7933335645f98a183efbbf472139125bd061