Re-create analytics.settings when users update data sharing option
Currently, AGP avoids to create analytics.settings if it doesn't exist
to avoid config cache miss in CI because analytics is generally not
needed in CI.

However, we need to ensure this analytics.settings file is created when
users update the data sharing setting. Otherwise, we cannot read whether
users optedIn and analytics is considered to be off.

Bug: n/a
Test: existing
Change-Id: I76103a6ccd5f4f2565be8b7066467f703610878b