[lint] Address test failure post-merge
After the merge of 2023.2.1 LintIdeGradleDetectorTest started failing
for Kotlin Script input, with each test file being annotated with a
number of errors essentially relating to the fact that the
Kotlin Script file did not have a script runtime attached.  This is in
somse sense expected, and indeed we already have to mask off the
extension that would normally prevent this file from having highlights
applied (in order that we can test the Lint highlights); mask away the
visitor which performs this highlighting.

(This is essentially a test-only problem; Lint does not rely on
resolving entities in its analysis of Kotlin Script files, so despite
the lack of a script runtime in the test the Lint detector is able to
operate successfully.  In production, we expect either there to be a
script runtime attached, or for its absence to be a bigger problem
than any Gradle-related highlighting.)

Fixes: 297433095
Test: this is it
Change-Id: Ia467af42bb4bc91de3b1db5ab2b86ec9cd6c08a0