Fix AndroidModularizeAction to be available for Kotlin
Its not clear why AndroidModularizeAction extended
BaseJavaRefactoringAction which is supposed to make actions only
available in Java files. It wasn't working as intended for a while,
but as of

https://github.com/JetBrains/intellij-community/commit/ede8bd8c13c6d45b6b011edcb450d8400768dc31

it correctly limits the availability of actions. So now the modularize action needs to extends BaseRefactoringAction in order to support
Kotlin.

Also, the module list is now sorted alphabetically.

Bug: 296548370
Test: manual
Change-Id: I463a66e5900c5df3ebd1c71bfff712de6e467fbe