Fix formatter used for inspection on layout xml
AndroidXmlFormattingModelBuilder should be used in all cases dealing
with formatting for Android layout xml files (as well as a few others).

In some cases, it's not correctly identifying layout files.
Specifically, this happens when the file passed in is a virtual copy of
a the original file, as is the case when running formatting to detect
changes but not actually make changes to the original file.

This change adds logic to detect such virtual files and make decisions
based upon the original file's type instead.

Fixes: 297790370
Test: AndroidXmlFormatterTest.testLayoutNoChangesNeeded
Change-Id: Id161e0a12210e6a84abae0fa62fbe9efbd7bc182