Remove NELE_USE_SHARED_ISSUE_PANEL_FOR_DESIGN_TOOLS flag
The NELE_USE_SHARED_ISSUE_PANEL_FOR_DESIGN_TOOLS flag has been enabled
for a while now and can be removed. This allows us to clean up a lot of
the code of the old issues panel, which is coupled with DesignSurface.

This change removes the flag and an initial chunk of the code that was
wrapped in it. More iterations should follow to clean up even more,
otherwise this change could become more complicated than it should and
we might miss something.

Bug: 298229332
Test: N/A
Change-Id: I2915071a8912260a24d58b95731e456d797e0b54