Fix the order of source folders in source type node
The order matters because some clients check only one
VIRTUAL_FILE and for this case it will be first file in the list.
one example of this is Navigation bar, it only shows the first file of
this list when selection AndroidSourceTypeNode.
Another example is NewAndroidComponentAction#actionPerformed where
it checks that VIRTUAL_FILE as target directory is main source set.

Sort the order of source set based on same key they are sorted in the
tree.

Bug: 296824880
Test: AndroidSourceTypeNodeTest
Change-Id: I14e91af68a69271c65d2e4716d8885343578f8e3
(cherry picked from commit e74dd522bfb5bb5b8a3d6c6c092af111a23529ae)