[BA] split out historical results from the latest one
We do not want to serialise everything we have in the current latest
result. We want to separate historical results visually anyway as we
don't want to allow any actions (like running jetifier or configuration
cache trial builds) on the historic results.

At this moment keeping historic and actual recent results in the same
object lead to the problems with data that is too complex to try to
serialise and restore. e.g. it affects refactoring of
Request.RequestData object.

We need to split these two entities to avoid problems with not all
data being available in historical results.

Bug: 298955958
Test: Existing
Change-Id: Idf155f42d128afa23f9117978c3ebfadf7e97f5a