Send all classes on first LE for a file
Due to name mangling differences between Kotlin compiler versions,
names of lambda classes may differ between Gradle and LE generate
classes.

This can lead to a scenario where Live Edit thinks a class is named
"Foo$1$lambda", but the version in the APK is named "Foo$1$1", or
something similar.

Live Edit cannot tell that anything has changed, since the compiler we
use to generate a diff "baseline" is our embedded compiler, not the
Gradle one; from our perspective, the class has always been named
"Foo$1$lambda".

To solve this, we simply send every class to the device the first time
we generate it, regardless if our diff detects a change. This ensures
that we never miss a lambda renaming between APK and LE.

Fixes: 297097789
Test: ComposableCompileTest.kt
Change-Id: Ie3d0cd13152aa0c332e25295b64dc1d47e7d39b1